Monotone-Parent: f4cb2b00dc9fb253acaaa97d083eec196e594fac Monotone-Revision: 698a166193d0842dd18a4a9d456862b37e990823
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2006-08-01T19:18:06
Monotone-Branch: ca.inverse.sogo

+++ UI/SOGoUI/UIxComponent.m
@@ -498,28 +498,33 @@ static BOOL uixDebugEnabled = NO;
   WOResourceManager *rm;
   NSBundle *pageBundle;
 
-  if (!pageToURL)
-    pageToURL = [[NSMutableDictionary alloc] initWithCapacity: 32];
-
-  url = [pageToURL objectForKey: filename];
-  if (!url)
+  if (filename)
     {
-      rm = [self pageResourceManager];
-      page = [[self context] page];
-      pageBundle = [NSBundle bundleForClass: [page class]];
-      url = [rm urlForResourceNamed: filename
-		inFramework: [pageBundle bundlePath]
-		languages: nil
-		request: [[self context] request]];
+      if (!pageToURL)
+        pageToURL = [[NSMutableDictionary alloc] initWithCapacity: 32];
+
+      url = [pageToURL objectForKey: filename];
       if (!url)
-	url = @"";
-      else
-        if ([url hasPrefix: @"http"])
-          url = [url hostlessURL];
-      [pageToURL setObject: url forKey: filename];
-    }
+        {
+          rm = [self pageResourceManager];
+          page = [[self context] page];
+          pageBundle = [NSBundle bundleForClass: [page class]];
+          url = [rm urlForResourceNamed: filename
+                    inFramework: [pageBundle bundlePath]
+                    languages: nil
+                    request: [[self context] request]];
+          if (!url)
+            url = @"";
+          else
+            if ([url hasPrefix: @"http"])
+              url = [url hostlessURL];
+          [pageToURL setObject: url forKey: filename];
+        }
 
 //   NSLog (@"url for '%@': '%@'", filename, url);
+    }
+  else
+    url = @"";
 
   return url;
 }
